Modified Home's header component
