+++ src/screens/home/Collectibles.tsx
@@ -1,11 +1,4 @@
-import React, {
-  useState,
-  useContext,
-  useEffect,
-  useMemo,
-  useRef,
-  useCallback,
-} from 'react';
+import React, { useState, useMemo } from 'react';
 import { useTheme } from 'react-native-paper';
 import { Platform, StyleSheet, View } from 'react-native';
 import { CommonActions, useNavigation } from '@react-navigation/native';
@@ -13,7 +6,6 @@ import { useQuery } from '@realm/react';
 import { useMutation } from 'react-query';
 
 import ScreenContainer from 'src/components/ScreenContainer';
-import { LocalizationContext } from 'src/contexts/LocalizationContext';
 import { NavigationRoutes } from 'src/navigation/NavigationRoutes';
 import CollectibleAssetsList from './components/CollectibleAssetsList';
 import HomeHeader from './components/HomeHeader';
@@ -22,8 +14,6 @@ import { hp } from 'src/constants/responsive';
 import { RealmSchema } from 'src/storage/enum';
 import useWallets from 'src/hooks/useWallets';
 import { ApiHandler } from 'src/services/handler/apiHandler';
-import useRgbWallets from 'src/hooks/useRgbWallets';
-import { AppContext } from 'src/contexts/AppContext';
 import {
   Asset,
   AssetFace,
@@ -32,23 +22,14 @@ import {
   Collectible,
   UniqueDigitalAsset,
 } from 'src/models/interfaces/RGBWallet';
-import AppType from 'src/models/enums/AppType';
-import CurrencyKind from 'src/models/enums/CurrencyKind';
 
 function Collectibles() {
   const theme: AppTheme = useTheme();
   const styles = useMemo(() => getStyles(theme), [theme]);
 
-  const { translations } = useContext(LocalizationContext);
-  const { common, sendScreen, home } = translations;
-
-  const app = useQuery(RealmSchema.TribeApp)[0];
   const navigation = useNavigation();
 
   const refreshRgbWallet = useMutation(ApiHandler.refreshRgbWallet);
-
-  const rgbWallet = useRgbWallets({}).wallets[0];
-  const { setAppType } = useContext(AppContext);
 
   const refreshWallet = useMutation(ApiHandler.refreshWallets);
   const wallet = useWallets({}).wallets[0];
@@ -64,8 +45,6 @@ function Collectibles() {
   );
 
   const [refreshing, setRefreshing] = useState(false);
-  const [image, setImage] = useState(null);
-  const [walletName, setWalletName] = useState(null);
 
   const assets: Asset[] = useMemo(() => {
     return [...collectibles, ...udas]
@@ -73,50 +52,8 @@ function Collectibles() {
       .sort((a, b) => b.timestamp - a.timestamp);
   }, [collectibles, udas]);
 
-  const prevBalances = useRef(null);
-
-  const calculateBalance = useCallback(() => {
-    if (app?.appType === AppType.NODE_CONNECT) {
-      return rgbWallet?.nodeBtcBalance?.vanilla?.spendable ?? 0;
-    }
-    return (
-      (wallet?.specs?.balances?.confirmed ?? 0) +
-      (wallet?.specs?.balances?.unconfirmed ?? 0)
-    );
-  }, [
-    app?.appType,
-    rgbWallet?.nodeBtcBalance?.vanilla?.spendable,
-    wallet?.specs?.balances?.confirmed,
-    wallet?.specs?.balances?.unconfirmed,
-  ]);
-
-  const balances = useMemo(() => {
-    const newBalance = calculateBalance();
-    if (prevBalances.current === newBalance) return prevBalances.current;
-    prevBalances.current = newBalance;
-    return newBalance;
-  }, [calculateBalance]);
-
-  const walletImage = useMemo(() => app?.walletImage, [app?.walletImage]);
-  const displayName = useMemo(() => app?.appName, [app?.appName]);
-
-  useEffect(() => {
-    setImage(walletImage);
-    setWalletName(displayName);
-  }, [walletImage, displayName]);
-
   const handleNavigation = (route, params?) => {
     navigation.dispatch(CommonActions.navigate(route, params));
-  };
-
-  const toggleDisplayMode = () => {
-    setAppType(
-      app.currencyMode === CurrencyKind.SATS
-        ? CurrencyKind.BITCOIN
-        : app.currencyMode === CurrencyKind.BITCOIN
-        ? CurrencyKind.FIAT
-        : CurrencyKind.SATS,
-    );
   };
 
   const handleRefresh = () => {
@@ -129,26 +66,7 @@ function Collectibles() {
   return (
     <ScreenContainer style={styles.container}>
       <View style={styles.headerWrapper}>
-        <HomeHeader
-          profile={image}
-          username={walletName}
-          balance={balances}
-          onPressScanner={() =>
-            handleNavigation(NavigationRoutes.SENDSCREEN, {
-              receiveData: 'send',
-              title: common.send,
-              subTitle: sendScreen.headerSubTitle,
-              wallet,
-            })
-          }
-          onPressNotification={() => console.log('Notification pressed')}
-          onPressProfile={() =>
-            handleNavigation(NavigationRoutes.WALLETDETAILS, {
-              autoRefresh: true,
-            })
-          }
-          onPressTotalAmt={toggleDisplayMode}
-        />
+        <HomeHeader />
       </View>
       <CollectibleAssetsList
         listData={assets}
