+++ src/screens/home/components/HomeHeader.tsx
@@ -1,15 +1,16 @@
-import * as React from 'react';
+import React, { useState, useEffect } from 'react';
 import { View, StyleSheet, Platform } from 'react-native';
 import { useTheme } from 'react-native-paper';
+import { useQuery } from '@realm/react';
+import { CommonActions, useNavigation } from '@react-navigation/native';
+import { useMMKVBoolean, useMMKVString } from 'react-native-mmkv';
 
 import AppText from 'src/components/AppText';
 import { hp, wp } from 'src/constants/responsive';
-
 import IconBitcoin from 'src/assets/images/icon_btc2.svg';
 import IconBitcoinLight from 'src/assets/images/icon_btc2_light.svg';
 import IconScanner from 'src/assets/images/icon_scanner.svg';
 import IconScannerLight from 'src/assets/images/icon_scanner_light.svg';
-// import IconNotification from 'src/assets/images/icon_notifications.svg';
 import IconWrapper from 'src/components/IconWrapper';
 import AppTouchable from 'src/components/AppTouchable';
 import { AppTheme } from 'src/theme';
@@ -17,48 +18,70 @@ import { LocalizationContext } from 'src/contexts/LocalizationContext';
 import HomeUserAvatar from './HomeUserAvatar';
 import useBalance from 'src/hooks/useBalance';
 import { Keys } from 'src/storage';
-import { useMMKVBoolean, useMMKVString } from 'react-native-mmkv';
 import CurrencyKind from 'src/models/enums/CurrencyKind';
+import useWallets from 'src/hooks/useWallets';
+import { NavigationRoutes } from 'src/navigation/NavigationRoutes';
+import { TribeApp } from 'src/models/interfaces/TribeApp';
+import { RealmSchema } from 'src/storage/enum';
+import AppType from 'src/models/enums/AppType';
+import useRgbWallets from 'src/hooks/useRgbWallets';
 
-type HomeHeaderProps = {
-  profile: string;
-  username: string;
-  balance: string;
-  onPressProfile: () => void;
-  onPressScanner: () => void;
-  onPressNotification: () => void;
-  onPressTotalAmt: () => void;
-};
-function HomeHeader(props: HomeHeaderProps) {
-  const {
-    profile,
-    username,
-    balance,
-    onPressScanner,
-    onPressNotification,
-    onPressProfile,
-    onPressTotalAmt,
-  } = props;
+function HomeHeader() {
   const theme: AppTheme = useTheme();
+  const navigation = useNavigation();
   const [isThemeDark] = useMMKVBoolean(Keys.THEME_MODE);
   const styles = React.useMemo(() => getStyles(theme), [theme]);
   const { translations } = React.useContext(LocalizationContext);
-  const { home } = translations;
+  const { home, common, sendScreen } = translations;
   const { getBalance, getCurrencyIcon } = useBalance();
   const [currentCurrencyMode] = useMMKVString(Keys.CURRENCY_MODE);
   const initialCurrencyMode = currentCurrencyMode || CurrencyKind.SATS;
+  const wallet = useWallets({}).wallets[0];
+  const app = useQuery<TribeApp>(RealmSchema.TribeApp)[0];
+  const rgbWallet = useRgbWallets({}).wallets[0];
+  const [image, setImage] = useState(null);
+  const [walletName, setWalletName] = useState(null);
+
+  useEffect(() => {
+    if (app?.walletImage || app?.appName) {
+      setImage(app.walletImage);
+      setWalletName(app.appName);
+    }
+  }, [app]);
+
+  const balances = React.useMemo(() => {
+    if (app.appType === AppType.NODE_CONNECT) {
+      return rgbWallet?.nodeBtcBalance?.vanilla?.spendable || '';
+    }
+    return wallet.specs.balances.confirmed + wallet.specs.balances.unconfirmed;
+  }, [
+    app.appType,
+    rgbWallet?.nodeBtcBalance?.vanilla?.spendable,
+    wallet?.specs.balances,
+  ]);
+
+  const handleNavigation = (route, params?) => {
+    navigation.dispatch(CommonActions.navigate(route, params));
+  };
+
   return (
     <View>
       <View style={styles.container}>
-        <AppTouchable onPress={onPressProfile} style={styles.contentWrapper}>
+        <AppTouchable
+          onPress={() => {
+            handleNavigation(NavigationRoutes.WALLETDETAILS, {
+              autoRefresh: true,
+            });
+          }}
+          style={styles.contentWrapper}>
           <View style={styles.contentWrapper}>
-            <HomeUserAvatar imageSource={profile} />
+            <HomeUserAvatar imageSource={image} />
             <View style={styles.userDetailsWrapper}>
               <AppText
                 numberOfLines={1}
                 variant="heading1"
                 style={styles.usernameText}>
-                {username ? username : 'Satoshi’s Palette'}
+                {walletName ? walletName : 'Satoshi’s Palette'}
               </AppText>
               <View style={styles.balanceWrapper}>
                 {initialCurrencyMode !== CurrencyKind.SATS &&
@@ -68,7 +91,7 @@ function HomeHeader(props: HomeHeaderProps) {
                     15,
                   )}
                 <AppText variant="body2" style={styles.balanceText}>
-                  &nbsp;{getBalance(balance)}
+                  &nbsp;{getBalance(balances)}
                 </AppText>
                 {initialCurrencyMode === CurrencyKind.SATS && (
                   <AppText variant="caption" style={styles.satsText}>
@@ -80,34 +103,19 @@ function HomeHeader(props: HomeHeaderProps) {
           </View>
         </AppTouchable>
         <View style={styles.iconWrapper}>
-          <IconWrapper onPress={onPressScanner}>
+          <IconWrapper
+            onPress={() => {
+              handleNavigation(NavigationRoutes.SENDSCREEN, {
+                receiveData: 'send',
+                title: common.send,
+                subTitle: sendScreen.headerSubTitle,
+                wallet,
+              });
+            }}>
             {isThemeDark ? <IconScanner /> : <IconScannerLight />}
           </IconWrapper>
-          {/* To Do */}
-          {/* <IconWrapper onPress={onPressNotification}>
-            <IconNotification />
-          </IconWrapper> */}
         </View>
       </View>
-      {/* <View style={styles.balanceContainer}>
-        <View>
-          <AppText variant="body2" style={styles.totalBalText}>
-            {home.totalBalance}
-          </AppText>
-        </View>
-        <AppTouchable style={styles.balanceWrapper} onPress={onPressTotalAmt}>
-          {initialCurrencyMode !== CurrencyKind.SATS &&
-            getCurrencyIcon(IconBitcoin, 'dark', 30)}
-          <AppText variant="pageTitle2" style={styles.balanceText}>
-            &nbsp;{getBalance(balance)}
-          </AppText>
-          {initialCurrencyMode === CurrencyKind.SATS && (
-            <AppText variant="caption" style={styles.satsText}>
-              sats
-            </AppText>
-          )}
-        </AppTouchable>
-      </View> */}
     </View>
   );
 }
@@ -151,18 +159,6 @@ const getStyles = (theme: AppTheme) =>
       // justifyContent: 'space-between',
       width: '33%',
       alignItems: 'flex-end',
-    },
-    iconTouchableArea: {
-      height: '60%',
-    },
-    balanceContainer: {
-      alignItems: 'center',
-      marginTop: hp(30),
-      marginBottom: hp(5),
-    },
-    totalBalText: {
-      color: theme.colors.secondaryHeadingColor,
-      fontWeight: '400',
     },
   });
 export default HomeHeader;
